pyproject.toml
[project]
name = "agentops"
version = "0.1.0"
description = "AgentOps — lightweight companion browser for local agent services"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "Steff Van Haverbeke" }]
keywords = ["ai-agents", "desktop", "webview", "localhost", "dev-tools"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: MacOS X :: Cocoa",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "pywebview>=5.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "typer>=0.12",
]

[project.optional-dependencies]
build = ["pyinstaller>=6.0"]

[project.scripts]
agentops = "agentops.cli:app"

[project.urls]
Homepage = "https://github.com/CoachSteff/agentops"
Repository = "https://github.com/CoachSteff/agentops"
Issues = "https://github.com/CoachSteff/agentops/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agentops"]
